Dragon Block Added (No Functionality Yet)

+++ src/main/java/net/mcreator/stupiddragonblockc/procedures/KatchinGetProcedure.java
@@ -28,7 +28,7 @@ public class KatchinGetProcedure {
 					_level.addFreshEntity(entityToSpawn);
 				}
 			} else if (EnchantmentHelper.getItemEnchantmentLevel(Enchantments.BLOCK_FORTUNE, (entity instanceof LivingEntity _livEnt ? _livEnt.getMainHandItem() : ItemStack.EMPTY)) == 1) {
-				for (int index0 = 0; index0 < (int) (Mth.nextInt(RandomSource.create(), 1, 2)); index0++) {
+				for (int index0 = 0; index0 < Mth.nextInt(RandomSource.create(), 1, 2); index0++) {
 					if (world instanceof Level _level && !_level.isClientSide()) {
 						ItemEntity entityToSpawn = new ItemEntity(_level, x, y, z, new ItemStack(StupidDbcModItems.RAW_KATCHIN.get()));
 						entityToSpawn.setPickUpDelay(10);
@@ -36,7 +36,7 @@ public class KatchinGetProcedure {
 					}
 				}
 			} else if (EnchantmentHelper.getItemEnchantmentLevel(Enchantments.BLOCK_FORTUNE, (entity instanceof LivingEntity _livEnt ? _livEnt.getMainHandItem() : ItemStack.EMPTY)) == 2) {
-				for (int index1 = 0; index1 < (int) (Mth.nextInt(RandomSource.create(), 1, 3)); index1++) {
+				for (int index1 = 0; index1 < Mth.nextInt(RandomSource.create(), 1, 3); index1++) {
 					if (world instanceof Level _level && !_level.isClientSide()) {
 						ItemEntity entityToSpawn = new ItemEntity(_level, x, y, z, new ItemStack(StupidDbcModItems.RAW_KATCHIN.get()));
 						entityToSpawn.setPickUpDelay(10);
@@ -44,7 +44,7 @@ public class KatchinGetProcedure {
 					}
 				}
 			} else if (EnchantmentHelper.getItemEnchantmentLevel(Enchantments.BLOCK_FORTUNE, (entity instanceof LivingEntity _livEnt ? _livEnt.getMainHandItem() : ItemStack.EMPTY)) == 3) {
-				for (int index2 = 0; index2 < (int) (Mth.nextInt(RandomSource.create(), 1, 4)); index2++) {
+				for (int index2 = 0; index2 < Mth.nextInt(RandomSource.create(), 1, 4); index2++) {
 					if (world instanceof Level _level && !_level.isClientSide()) {
 						ItemEntity entityToSpawn = new ItemEntity(_level, x, y, z, new ItemStack(StupidDbcModItems.RAW_KATCHIN.get()));
 						entityToSpawn.setPickUpDelay(10);
